+++ src/Text/Show/Text/System.hs
@@ -11,13 +11,14 @@ Imports 'Show' instances for @System@ modules.
 -}
 module Text.Show.Text.System () where
 
-import Text.Show.Text.System.Locale           ()
-import Text.Show.Text.System.Random           ()
-import Text.Show.Text.System.Time             ()
+import Text.Show.Text.System.Console.Haskeline ()
+import Text.Show.Text.System.Locale            ()
+import Text.Show.Text.System.Random            ()
+import Text.Show.Text.System.Time              ()
 
 #if defined(mingw32_HOST_OS)
-import Text.Show.Text.System.Win32            ()
+import Text.Show.Text.System.Win32             ()
 #else
-import Text.Show.Text.System.Console.Terminfo ()
-import Text.Show.Text.System.Posix            ()
+import Text.Show.Text.System.Console.Terminfo  ()
+import Text.Show.Text.System.Posix             ()
 #endif