+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.2
 * Exposed `showbKindPrec` and `showbPredPrec` with later versions of Template Haskell
-* Added instances for the `binary`, `hoopl`, `terminfo`, and `utf8-string` libraries
+* Added instances for the `binary`, `haskeline`, `hoopl`, `terminfo`, and `utf8-string` libraries
 
 # 0.1
 * Initial commit